Split ObjectiveForm component
